requirements.txt
db-client @ git+https://github.com/vvSSpace/db_client.git@2886ee971fa8e7dc9c8eb35979eabdc6c8448493
dm-api-account @ git+https://github.com/vvSSpace/dm_api_account.git@e3dfbd62015e1857b5d150a8258491a9a6b9aba8
orm-client @ git+https://github.com/vvSSpace/orm_client.git@53add86fbeca8dc35d0261c3001ca46065e92c28
restclient @ git+https://github.com/vvSSpace/restclient.git@cf0c11573d069219cc55c332a8cc679e82c0c9eb
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.5.0
anyio==3.7.1
argcomplete==3.1.1
attrs==23.1.0
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
curlify==2.2.1
datamodel-code-generator==0.21.1
distconfig3==1.0.1
dnspython==2.3.0
docopt==0.6.2
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
genson==1.2.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflect==5.6.2
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.6
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.2
openpyxl==2.4.11
packaging==23.1
pathable==0.4.3
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
prance==23.6.21.0
psycopg2==2.9.6
pydantic==2.0
pydantic_core==2.0.1
PyHamcrest==2.0.4
pyrsistent==0.19.3
PySnooper==1.1.1
pytest==7.3.2
PyYAML==6.0
records==0.5.3
requests==2.31.0
rfc3339-validator==0.1.4
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.46
structlog==23.1.0
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
vyper-config==1.1.1
watchdog==3.0.0
Werkzeug==2.3.6
grpcio-tools==1.56.2
betterproto==2.0.0-beta5
